requirements.txt
selenium==4.10.0
beautifulsoup4==4.12.2
openai==0.28.0
langchain==0.0.304
nltk==3.8.1
spacy==3.6.0
transformers==4.33.2
pymongo==4.5.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.4
tqdm==4.66.1
loguru==0.7.0
retrying==1.3.4
python-dotenv==1.0.0
